+++ fltk-sys/bind.sh
@@ -17,3 +17,4 @@ bindgen --use-core --ctypes-prefix libc fltk-sys/cfltk/include/cfl_table.h -o fl
 bindgen --use-core --ctypes-prefix libc fltk-sys/cfltk/include/cfl_tree.h -o fltk-sys/src/tree.rs
 bindgen --use-core --ctypes-prefix libc fltk-sys/cfltk/include/cfl.h -o fltk-sys/src/fl.rs
 bindgen --use-core --ctypes-prefix libc fltk-sys/cfltk/include/cfl_surface.h -o fltk-sys/src/surface.rs
+bindgen --use-core --ctypes-prefix libc fltk-sys/cfltk/include/cfl_printer.h -o fltk-sys/src/printer.rs