+++ fltk-sys/src/lib.rs
@@ -22,6 +22,7 @@ pub mod valuator;
 pub mod widget;
 pub mod window;
 pub mod surface;
+pub mod printer;
 
 #[cfg(test)]
 mod tests {